Removed shadow_paradigm option
On second thought, the "shadow paradigm" is not really the appropriate way to handle the problem.  The set methods on the controller should be smart enough to copy over the value to the controller if needed.  This is a refactor of the code that, for the most part, removes the dependence on the shadow magic.