Positive behavior circuit now obeys zBUS trigger, other misc cleanup
For the positive behavior circuit, the zBUS trigger must be set to high to
initiate data collection.  This allows the circuit to be loaded and started by
the DSPProcess.  We can then set zBUS trigger A to high to start the experiment.
Prior to this, the circuit automatically started collecting data as soon as the
circuit started.

Eventually, we need to switch all the aversive and physiology circuits over to
this approach as well.

Also includes some minor bugfixes and cleanup to the positive DT code.