Updated the channel module to use a FileMixin for saving data to HDF5 file
Both Timeseries and Channel classes can now "mixin" a class, FileMixin, that
handles most of the infrastructure of saving data to a HDF5 file.  To use this
mixin class, be sure all important attributes of the class (e.g. fs and t0) have
the 'attr' metadata set to True when defining the trait.

Also, the FileMixin class exposes the data source via the _buffer attribute.