Hand-edited "merge" to bring in shock controller work from a closed branch
An earlier attempt to do a Hg-managed merge of a branch with the shock
controller failed.  This is probably due to the fact that the branches had
diverged so much by the time I got around to attempting the merge.  Hence, I
decided it was simpler to do the merge by hand.

This work allows the user to control the shock level programmatically via the
GUI and will save the current shock level to the experiment trial log.