Fixed test_experiment to overwrite temporary file if it is present
Sometimes the program does not exit cleanly on launch, leaving a temporary file
that is not writeable.  By default, Neurobehavior attempts to append to a file
if one is specified.  However, it is safe to overwrite temporary files so we
open in write mode that way we don't have to deal with corrupted remnants of
temporary files from crashes.