Modified the zoom behavior for the ChannelRangeTool
Changed control+scroll so that scrolling up zooms in along the index axis and
scrolling down zooms out along the index axis.  This is now consistent with the
behavior of the regular wheel where normal wheel up zooms in along the value
axis and normal wheel down zooms out along the value axis.