Bugfix to ensure that hardware attenuation is set properly on experiment start
The new attenuation values are compared with the current_hw_attX attributes to
determine whether any work needs to be done (i.e. uploading the new value to the
RZ6, updating the neurogen blocks, etc).  This was getting skipped.  Changed
the current_hw_attX attributes to -1 to indicate the attenuation has not been
initialized yet.