Fix to the code that applies changes to the context
The code was attempting to find the setter functions (e.g. set_parameter_name)
by simply doing a getattr.  If no setter is avialable, the resulting
AttributeError was captured and silently discarded.  However, if the setter
did exist and raised an AttributeError, this error was also getting silently
discarded.

To allow these AttributeErrors to bubble up the hierarchy, I now check to ensure
that the setter exists before attempting to call it, that way I know all
AttributeErrors are due to an actual bug in the setter function itself.