Bugfix to facilitate loading cached/saved paradigms from disk
The constant limits paradigm used to expect the trial setting list to be in a
very specific order; however, when the paradigm was loaded from disk (i.e. from
a cached copy or from a saved *.par file), the list was sometimes loaded out
of order.  To workaround this issue, I have modified the paradigm mixin to
scan the trial_setting list, remove duplicate reminds and nogos and ensure there
is at least one remind, one nogo and one go in the list.  The user is expected
to visually confirm the contents of the list before starting the experiment.

There is no reason why one can't have multiple nogos/reminds; however, this
requires implementing the corresponding logic in the controller to select from
a sequence of nogos/reminds.  That exercise will be left to the person who
actually needs this logic.