Increased size of buffers in circuit
The default settings for the buffer size could only spool ~5 seconds of data
before overflowing.  To ensure no data was lost, Python had to download data
continuously.  Once in a while, the software would freeze for a duration
> 5 seconds, resulting in loss of data.  The new settings for buffer size can
hold a *minimum* of 50 seconds of data before overflowing.

* TTL can hold 65 seconds
* craw can hold 56 seconds
* cfilt can hold 51 seconds

We are pretty much at the limit of the memory we can utilize in the RZ5 without
adding a second CPU.  Either case, a minimum of 51 seconds is sufficient.  If
the program hangs for more than a couple of seconds, the experiment should be
terminated anyway.