Bugfixes to appetitive code
Note that trial_duration is essentially determined by the duration of the token
plus several other factors (e.g. reaction and response time) and is a misnomer.
What we *actually* used trial_duration for was to determine the duration of the
token.  I've renamed this to duration and each signal mixin class is responsible
for handling the duration of their own token.