Alpha version of program to review physiology data and extract spikes
* Allows the user to select bad channels on the fly and automatically updates
the referencing configuration to exclude these bad channels. Referencing uses
to the "all" approach (e.g. subtracting the average of the "good" channels from
each channel). * Allows the user to change filter settings on the fly,
automatically updating the GUI to reflect the new filter settings. Right now
only the low and high pass frequencies can be changed. Note that the GUI
updates as you type the new value into the field (that's not a good thing since
it's computationally expensive to recompute the filter coefficients and
refilter the signal -- ideally there should be an "apply" button for this
specific field). * Once the user is happy with the filtering and referencing,
they may select several channels to export to KlustaKwik format (exported
channels are those checked off in the "Ex?" column). The filename is "test.txt"
* Alternatively, they may choose to extract only the waveforms. Waveform
extraction is a three-step process 1) select File -> Compute noise floor 2)
Select the channels to extract by checking off the corresponding box in "Ex?"
column 3) Select File -> threshold spikes Data will be saved to a HDF5 file
called "test.hd5". This is a very simple format that should be easy for
Matlab's HDF5 routines to handle. Some additional data will need to be provided
to Chronux to help with spike-sorting but this should easily be figured out.
There is a way to hack the HDF5 file to trick Matlab into thinking it is a
native *.mat file; however, I don't have the time to work on this. * A tab
allows you to click on a specific trial to jump to that point in the file.

Note that some of the user feedback is printed to the console, so when you run
long operations (e.g. thresholding, noise floor computation, data saving), be
sure to look at the console so you can monitor progress.

Eventually the GUI can hopefully be refined to allow the user to load new
physiology datasets, specify the output filenames and customize some of the
data extraction options (e.g. the number of samples in the waveform to extract,
the threshold value).

- For efficiency and memory purposes, all processing is performed in chunks of
~10 MB (this translates to ~1.5 million samples from each channel). Playing
with the chunk size (by editing the CHUNK_SIZE variable) may speed up data
processing.

- As with the on-time data visualization, the off-line data visualization uses
a very aggressive caching strategy to optimize performance, so you may notice
the GUI is not properly updating based on changes you make. These are
considered bugs and you should make a list of them.

- Since we process the code in chunks, there will be some boundary effects at
the end of each chunk. For example, most filtering algorithms pad the edges of
the signal with some zeros. Instead, we should be padding the edges of the
signal with data from adjacent chunks prior to filtering. Likewise, I would
like to implement forward/reverse filtering to eliminate phase shifts. For now,
only a single-pass forward filter is implemented because the two-pass filter
does not support 2D arrays.

- Spikes at the very edges of each chunk may get discarded if they cross a
chunk boundary. Eventually we will need to add some code to pull in the data
from the adjacent chunk for processing.

- This algorithm uses only the first 10 MB of the dataset to compute the
standard deviation of the noise for each channel. The algorithm I use for
computing standard deviation uses the median. Standard algorithms for computing
the median value of a dataset require loading the full dataset into memory.
This is not feasible with some of the larger physiology datasets; hence, I only
load the first "chunk" into memory for computing the noise floor. There do
exist some algorithms for calculating the median incrementally, allowing us to
break up the dataset into more manageable chunks and update our estimate of the
median after each chunk; however, I am not going to implement one of these
algorithms since they are relatively complex.

- This code uses the same GUI widgets as Neurobehavior, which is why it is part
of the Neurobehavior package rather than a stand-alone script. In theory it
should run on any platform, but it really depends on whether you can get the
appropriate libraries installed. It will certainly run on a PC or Linux.

- I had to fix some bugs and modify the code for the GUI widgets. At some point
we should merge these changes back into the main Neurobehavior code. This will
require some testing to make sure I haven't broken any behavior that the
experiment code relies on.