UltraMegaSort2000 should not be in the repository
This was added by a student because we modified some files that UMS2000 uses.
However, since we don't own the copyright on UMS2000, we should not be including
the code in our repository.  What we can do, instead, is include only the files
we have modified.  By virtue of Matlab's lack of namespaces, we just need to
make sure that the modified files are at the top of Matlab's path.  The two
files in question are reintegrate_outliers.m and remove_outliers.m.

Just make sure that the path to neurobehavior/matlab has greater precedence than
the path to the local copy of UltraMegaSort2000.