Added metadata to running RMS that indicates exit status of function
If the function is aborted before it is able to process the full waveform, then
we need to know this.  A new attribute is added to the output node called
'aborted'.  If the value of the attribute is True (or the attribute is missing),
this means that the function did not successfully process the full waveform.