Updated the running rms algorithm to better handle the last chunk
In the past, the last chunk was discarded because it needed special handling
when reshaping.  In some cases, the last chunk included a few important regions
of data, so I went back to better handle the processing of the last chunk.  Now,
all but the very last few samples (the last few samples always will be < window_step_samples) are processed.