Added metadata to mark unique versions of the analyzed data
Sometimes we go back and tweak the spike extraction, filtering and censor
settings.  To ensure that "downstream" files (e.g. the sorted spiketimes file)
are based on the most recent version of the extracted file, I've added some
additional metdata.  The UUID hash is globally unique and will change every
time the extracted spiketimes file is reprocessed.  Likewise,
the last_<operation> filed will store the time (in seconds since the Unix epoch)
that the operation was last performed.  By comparing this information with the
information stored in the sorted spiketimes file, you can determine whether
the sorted spiketimes are based on the most recent version of the extracted
spiketimes file.