Ensure changes to HDF5 file in review physiology GUI are saved immediately
PyTables does not always save changes to the file on disk until the file is
actually closed.  Consequently, if the file is left open in the GUI and the
user runs extract_spikes.py on the file, the latest changes to the extract
settings may not be seen by the extract_spikes.py script.