ENH: Removed NYU/Sanes-specific code from cns.analysis
Most of the code in cns.analysis is very generic and can be used to process
multiple file formats that support a certain interface (e.g. see cns.channel).
To facilitate re-use of this code, removed some aspects that were very specific
to the NYU/Sanes workflow.  The primary entry-point (e.g. the analysis scripts
in the scripts folder) should be responsible for handling workflow issues that
are specific to a particular lab or group.