[ENH] Broke out apply/revert start/stop logic of AbstractExperimentController
The AbstractExperimentController makes many assumptions about the experiment
(e.g. type of hardware, file formats, etc.) that make it insufficient for using
in other labs.  However, some of the core functions (e.g. the apply/revert and
expression logic) are sufficiently useful for any scientific experiment that I
am going to start creating a mix-in class that is much more abstract (and better
suited for reuse).  The current mix-in class defined in this commit has
successfully been tested with National Instrument's hardware (a controller
inheriting from this mixin still needs to define the appropriate clauses for
initializing the experiment and polling the hardware).